requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
requests==2.31.0
mysql-connector-python==8.1.0
pyautogui==0.9.54
pyperclip==1.8.2